Remove capture of Context by BackendUtils
